perf(xabe-engine): store the KV cache in binary16
The cache was fp32 and everything that reads it paid for that twice: in
DRAM traffic, which is what attention spends most of its time on, and in
the tensor-core kernel's staging, which had to convert every element with
a three-instruction cvt/cvt/mov because Turing has no cvt.rn.f16x2.f32
(that is sm_80, checked rather than assumed).

Now it is binary16, which is what llama.cpp's cache holds and what
xabe-cache's planner has always assumed with DEFAULT_ELEM_SIZE = 2. Three
things follow:

  - 20 KiB per token instead of 40. At 131,072 positions the sequence
    state falls from 5.06 to 2.53 GiB.
  - Half the DRAM traffic for every attention kernel.
  - The tensor-core kernel stages K with **no conversion at all**: two
    adjacent dimensions of one key are already the packed B operand, so
    it is one aligned 32-bit load. V still needs assembling, because its
    two halves are two different keys, but from binary16 loads rather
    than from floats needing a convert each.

Interleaved at 8,192 tokens in 512-token chunks, winning both pairs:
1759.5 -> 1934.2 tok/s, +9.9%. That is 0.978x of llama.cpp on the same
card, from 0.784x before this session's attention work.

xabe-kernels gains an f16 module with the host side of the same
conversion, round-to-nearest-even, subnormals and overflow included. It
exists so the differential tests can round K and V *at the point they are
generated* and hand the identical values to both the device and the scalar
reference. That makes the comparison measure the kernel instead of the
cache's dtype -- without it the two differ by an operand rounding before
the kernel does anything, and the only way to pass would be a tolerance
wide enough to stop measuring. Its own unit tests pin the tie-breaking
direction, the half-ulp bound over 200,000 samples, saturation to infinity
rather than wraparound, and the subnormal round trip.

One structural test needed its anchor moved: it located the t1 value fold
by the literal "a += w_sh[w] * v[", which acquired an h2f(...) and wrapped.
It anchors on the prefix now, since the property is the order of the fold
against the rescale and not how the value is read.

+++ crates/xabe-engine/src/block/attention.rs
@@ -511,8 +511,8 @@ impl AttentionKernelSet {
 /// [`GdnState`](crate::block::gdn::GdnState) holds and why this model's cache does not grow the way a
 /// forty-layer dense model's would.
 pub struct KvCache {
-    k: CudaSlice<f32>,
-    v: CudaSlice<f32>,
+    k: CudaSlice<u16>,
+    v: CudaSlice<u16>,
     kv_dim: usize,
     max_seq: usize,
 }
@@ -531,8 +531,8 @@ impl KvCache {
     ) -> Result<Self, AttentionBlockError> {
         let kv_dim = config.attention.kv_heads as usize * config.attention.head_dim as usize;
         Ok(Self {
-            k: stream.alloc_zeros::<f32>(max_seq * kv_dim)?,
-            v: stream.alloc_zeros::<f32>(max_seq * kv_dim)?,
+            k: stream.alloc_zeros::<u16>(max_seq * kv_dim)?,
+            v: stream.alloc_zeros::<u16>(max_seq * kv_dim)?,
             kv_dim,
             max_seq,
         })
@@ -544,18 +544,24 @@ impl KvCache {
     }
 
     /// Device bytes held, both halves.
+    ///
+    /// binary16, which is 20 KiB per token at this model's geometry rather than
+    /// 40 — 2.53 GiB at 131,072 positions instead of 5.06. It is also the dtype
+    /// llama.cpp's cache uses, and the one `xabe-cache`'s planner has always
+    /// assumed (`DEFAULT_ELEM_SIZE = 2`), which this makes true rather than
+    /// aspirational.
     pub fn bytes(&self) -> u64 {
-        2 * (self.max_seq * self.kv_dim * size_of::<f32>()) as u64
+        2 * (self.max_seq * self.kv_dim * size_of::<u16>()) as u64
     }
 
     /// The cached keys, `[max_seq][kv_heads][head_dim]`, rotary already
     /// applied. Positions at or above the sequence length are zero.
-    pub fn keys(&self) -> &CudaSlice<f32> {
+    pub fn keys(&self) -> &CudaSlice<u16> {
         &self.k
     }
 
     /// The cached values, same layout, not normed and not rotated.
-    pub fn values(&self) -> &CudaSlice<f32> {
+    pub fn values(&self) -> &CudaSlice<u16> {
         &self.v
     }
 }
